fix warnings

+++ src/utils_rmw.c
@@ -169,7 +169,7 @@ user_verify (void)
 
 /* used to prevent a TOCTOU race condtion */
 bool
-is_modified (const char* file, const int dev, const int inode)
+is_modified (const char* file, const unsigned long int dev, const unsigned long int inode)
 {
   struct stat st;
   if (lstat (file, &st))
